Put the mixer in three places, one per layer
The orchestrator, its state and the frame-timing headers were split between
src/ and src/mixer/; they are all in src/mixer/ now. The mixer control
protocol helpers (mixer_command, status and scene parsing) were in
demos/mixer/control.py, imported through a sys.path shim by the TUI, the web
UI, the smoke test and two hardware checks; they belong with the connection
that carries them, so they are in avpmixer/control.py.

What is left is one rule: src/mixer is the engine, avpmixer is the library
demos/playlist also builds on, demos/mixer is this demo and nothing else.

Verified by rebuilding the native module (T4 build container, exit 0) and
running it: five wipes cached, preheat complete, a media wipe and a fade taken
on air, 18% GPU.
