+++ demos/mixer/smoke_test.py
@@ -9,18 +9,7 @@ import time
 
 from avpmixer.control import AvpConnection
 
-try:
-    from .control import (
-        mixer_command,
-        parse_mixer_status,
-        parse_scene_list,
-    )
-except ImportError:
-    from control import (  # type: ignore[no-redef]
-        mixer_command,
-        parse_mixer_status,
-        parse_scene_list,
-    )
+from avpmixer.control import mixer_command, parse_mixer_status, parse_scene_list
 
 
 async def _command(
