+++ src/mixer/MixerState.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#include "instance_shared.hpp"
-#include "util.hpp"
+#include "../instance_shared.hpp"
+#include "../util.hpp"
 #include <cstdint>
 #include <mutex>
 #include <unordered_map>
